pyproject.toml
[project]
name = "skin-cancer-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.46.1",
    "kaggle>=1.7.4.5",
    "kagglehub>=0.3.13",
    "keras>=3.11.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "streamlit>=1.49.1",
    "tensorflow>=2.20.0",
    "timm>=1.0.22",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
